requirements.txt
streamlit==1.28.1
pandas==2.3.3
numpy==1.24.3
openpyxl==3.1.5
plotly==6.3.1
google-auth-oauthlib==1.3.1
google-auth-httplib2==0.3.1
google-api-python-client==2.193.0
python-dotenv==1.0.0
pillow==10.0.0
kaleido==0.2.1
openai==1.3.0
